+++ src/Figures/ISquare.php
@@ -4,5 +4,5 @@ namespace Devavi\Architecture\Figures;
 
 interface ISquare
 {
-    public function squareArea(int $sideSquare): int;
+    public function squareArea(float $sideSquare): float;
 }
